requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
jinja2>=3.1.2
qdrant-client>=1.13.3
sentence-transformers>=2.2.2
datasets>=2.12.0
pandas>=2.0.0
python-dotenv>=1.1.0
httpx>=0.24.0
